test(update): move namespaced-path regression into its own suite under the max-lines limit
